requirements.txt
causalnex==0.12.1
community==1.0.0b1
graphviz==0.20.1
ipython==8.26.0
matplotlib==3.7.1
networkx==3.1
nltk==3.8.1
numpy==1.22.1
pandas==1.5.3
plotly==5.22.0
powerlaw==1.5
praw==7.7.0
python_igraph==0.11.5
pyvis==0.3.2
replicate==0.23.1
scikit_learn==1.5.1
scipy==1.7.3
seaborn==0.13.2
tqdm==4.66.1
uunet==1.1.4